fix(`weibo-bot`): use transient tabs to avoid memory leaks
There is a potential memory leak issue in Weibo home page or Google Chrome. Using singleton tabs was a good idea, which means less network usage, because reloading the website is needless. But for the long-term services, those advantages are nothing comparing to the memory management, if you do not expect your server to crash with out of memory error. So using transient tabs is a better idea, this can free unused memory correctly :)