fix(`weibo-bot`): clear textarea correctly
In the Weibo home page loading period, `text_textarea.clear()` does not seem to be effective. If there is an exception occurred in the previous job, the text to be sent in the current job would join the previous, which is not expected. Instead, using keyboard simulation `Ctrl + A` and `Del` should fix this bug :)